// frontend/src/api/generated/model/automationCapability.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SmartGrao.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */

export type AutomationCapability = typeof AutomationCapability[keyof typeof AutomationCapability];


export const AutomationCapability = {
  ManualRecord: 'ManualRecord',
  UnderValidation: 'UnderValidation',
  AutomationEnabled: 'AutomationEnabled',
} as const;