// frontend/src/api/generated/model/plantOrgan.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SmartGrao.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */

export type PlantOrgan = typeof PlantOrgan[keyof typeof PlantOrgan];


export const PlantOrgan = {
  Undefined: 'Undefined',
  Whorl: 'Whorl',
  Leaf: 'Leaf',
  EarLeaf: 'EarLeaf',
  StalkBase: 'StalkBase',
  Stalk: 'Stalk',
  Root: 'Root',
  Ear: 'Ear',
  Tassel: 'Tassel',
  Silk: 'Silk',
  Seedling: 'Seedling',
  WholePlant: 'WholePlant',
} as const;