// frontend/src/api/generated/model/protocolStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SmartGrao.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */

export type ProtocolStatus = typeof ProtocolStatus[keyof typeof ProtocolStatus];


export const ProtocolStatus = {
  Draft: 'Draft',
  Published: 'Published',
  Retired: 'Retired',
} as const;